fix s3 delete_many (#38)
* fix s3 delete_many

Couldn't find any documentation around this issue, but I empirically discovered that `delete_objects` with an empty array causes the request to fail. Also, this is a valid optimization anyway.

* add tests and comments

* add comment

* use mock

* install mock unconditionally

+++ flask_annex/s3.py
@@ -40,11 +40,16 @@ class S3Annex(AnnexBase):
         self._client.delete_object(Bucket=self._bucket_name, Key=key)
 
     def delete_many(self, keys):
+        # casting to tuple because keys might be iterable
+        objects = tuple({'Key': key} for key in keys)
+        if not objects:
+            # this is not just an optimization, boto fails if the array
+            # is empty
+            return
+
         self._client.delete_objects(
             Bucket=self._bucket_name,
-            Delete={
-                'Objects': tuple({'Key': key} for key in keys),
-            },
+            Delete={'Objects': objects},
         )
 
     def get_file(self, key, out_file):
